requirements.txt
beautifulsoup4
requests
googlesearch-python
languagetool-python
ollama
textblob